Factor out table display
